Webpack 5 doesn't include the assert package.
When building with ng build the following error is shown.

./node_modules/console-browserify/index.js:4:13-30 - Error: Module not found: Error: Can't resolve 'assert' in '/home/missy/github/hutcheb/incubator-streampipes/ui/node_modules/console-browserify'

BREAKING CHANGE: webpack < 5 used to include polyfills for node.js core modules by default.
This is no longer the case. Verify if you need this module and configure a polyfill for it.

If you want to include a polyfill, you need to:
	- add a fallback 'resolve.fallback: { "assert": require.resolve("assert/") }'
	- install 'assert'
If you don't want to include a polyfill, you can use an empty module like this:
	resolve.fallback: { "assert": false }

This seems to be hidden when running ng build --configuration production