docs: fix capitalization, definition of start/end_date in data_lake_measure.py
The definitions of start- and end_date are reversed:

end_date: This option restricts the queried data to be younger than the specified time. This means that only data after the specified timestamp will be retrieved. For example, if end_date is set to January 1, 2023, only data from January 2, 2023, or later would be retrieved, going forward in time.

start_date: This option restricts the queried data to be older than the specified time. This means that only data before the specified timestamp will be retrieved. If start_date is set to January 1, 2023, only data before January 1, 2023, would be retrieved, i.e., December 31st, 2022, and earlier, going back in time.

These definitions imply that the end date lies further in the past than the start_date. This is in fact wrong because it should be the other way around. Accordingly, the definitions have been corrected.